+++ pdfalyzer/binary/binary_scanner.py
@@ -10,22 +10,24 @@ from typing import Any, Iterator, Pattern
 from rich.panel import Panel
 from rich.table import Table
 from rich.text import Text
+from yaralyzer.bytes_match import BytesMatch
+from yaralyzer.config import YaralyzerConfig
+from yaralyzer.decoding.bytes_decoder import BytesDecoder
+from yaralyzer.encoding_detection.character_encodings import BOMS
+from yaralyzer.helpers.bytes_helper import print_bytes
+from yaralyzer.helpers.rich_text_helper import (CENTER, console, console_width, na_txt, prefix_with_plain_text_obj,
+     subheading_width)
+from yaralyzer.helpers.string_helper import escape_yara_pattern
+from yaralyzer.output.regex_match_metrics import RegexMatchMetrics
+from yaralyzer.yaralyzer import Yaralyzer
+from yaralyzer.util.logging import log
 
-from pdfalyzer.binary.bytes_decoder import BytesDecoder
-from pdfalyzer.binary.bytes_match import BytesMatch
 from pdfalyzer.config import PdfalyzerConfig
-from pdfalyzer.detection.constants.binary_regexes import QUOTE_REGEXES
-from pdfalyzer.detection.constants.character_encodings import BOMS
-from pdfalyzer.detection.constants.binary_regexes import DANGEROUS_INSTRUCTIONS
-from pdfalyzer.detection.regex_match_metrics import RegexMatchMetrics
-from pdfalyzer.detection.yara_scanner import YaraScanner
-from pdfalyzer.helpers.bytes_helper import clean_byte_string, print_bytes
-from pdfalyzer.helpers.rich_text_helper import (CENTER, DANGER_HEADER, NOT_FOUND_MSG, console, console_width,
-     generate_subtable, get_label_style, na_txt, pad_header, prefix_with_plain_text_obj, subheading_width)
+from pdfalyzer.detection.constants.binary_regexes import DANGEROUS_STRINGS, QUOTE_REGEXES, QUOTE_PATTERNS
+from pdfalyzer.helpers.rich_text_helper import (DANGER_HEADER, NOT_FOUND_MSG,
+     generate_subtable, get_label_style, pad_header)
 from pdfalyzer.helpers.string_helper import generate_hyphen_line, print_section_header
 from pdfalyzer.util.adobe_strings import CURRENTFILE_EEXEC
-from pdfalyzer.util.logging import log
-
 
 # For rainbow colors
 CHAR_ENCODING_1ST_COLOR_NUMBER = 203
@@ -44,31 +46,28 @@ class BinaryScanner:
         self.stream_length = len(_bytes)
         self.regex_extraction_stats = defaultdict(lambda: RegexMatchMetrics())
         self.suppression_notice_queue = []
-        self.yara_scanner = YaraScanner(_bytes, label)
 
     def check_for_dangerous_instructions(self) -> None:
         """Scan for all the strings in DANGEROUS_INSTRUCTIONS list and decode bytes around them"""
         print_section_header("Scanning Font Binary For Anything 'Mad Sus'...", style=DANGER_HEADER)
 
-        for instruction in DANGEROUS_INSTRUCTIONS:
-            instruction_regex = re.compile(re.escape(instruction), re.DOTALL)
-            label = f"({BOMS[instruction]}) " if instruction in BOMS else clean_byte_string(instruction)
-            self._process_regex_matches(instruction_regex, label, force=True)
+        for instruction in DANGEROUS_STRINGS:
+            label = f"({BOMS[instruction]}) " if instruction in BOMS else instruction
+            self._process_yara_matches(instruction, label, force=True)
 
     def force_decode_all_quoted_bytes(self) -> None:
         """Find all strings matching QUOTE_REGEXES (AKA between quote chars) and decode them with various encodings"""
-        quote_types = QUOTE_REGEXES.keys() if PdfalyzerConfig.QUOTE_TYPE is None else [PdfalyzerConfig.QUOTE_TYPE]
+        quote_types = QUOTE_PATTERNS.keys() if PdfalyzerConfig.QUOTE_TYPE is None else [PdfalyzerConfig.QUOTE_TYPE]
 
         for quote_type in quote_types:
-            quote_regex = QUOTE_REGEXES[quote_type]
+            quote_regex = QUOTE_PATTERNS[quote_type]
             print_section_header(f"Forcing Decode of {quote_type.capitalize()} Quoted Strings", style='color(100)')
-            self._process_regex_matches(quote_regex, label=f"{quote_type} quoted")
+            self._process_yara_matches(quote_regex, rules_label=f"{quote_type} quoted")
 
     def extract_regex_capture_bytes(self, regex: Pattern[bytes]) -> Iterator[BytesMatch]:
         """Finds all matches of regex_with_one_capture in self.bytes and calls yield() with BytesMatch tuples"""
         for i, match in enumerate(regex.finditer(self.bytes, self._eexec_idx())):
             yield(BytesMatch.from_regex_match(self.bytes, match, i + 1))
-
 
     # -------------------------------------------------------------------------------
     # These extraction iterators will iterate over all matches for a specific pattern.
@@ -85,7 +84,6 @@ class BinaryScanner:
     def extract_front_slash_quoted_bytes(self) -> Iterator[BytesMatch]:
         """Returns an interator over all strings surrounded by front_slashes (hint: regular expressions)"""
         return self.extract_regex_capture_bytes(QUOTE_REGEXES['front_slaash'])
-
 
     def print_stream_preview(self, num_bytes=None, title_suffix=None) -> None:
         """Print a preview showing the beginning and end of the stream data"""
@@ -150,31 +148,29 @@ class BinaryScanner:
         """Get the bytes after the 'eexec' demarcation line (if it appears). See Adobe docs for details."""
         return self.bytes.split(CURRENTFILE_EEXEC)[1] if CURRENTFILE_EEXEC in self.bytes else self.bytes
 
-    def _process_regex_matches(self, regex: Pattern[bytes], label: str, force: bool=False) -> None:
+    def _process_yara_matches(self, pattern: str, rules_label: str, force: bool = False) -> None:
         """Decide whether to attempt to decode the matched bytes, track stats. force param ignores min/max length"""
-        for bytes_match in self.extract_regex_capture_bytes(regex):
-            self.regex_extraction_stats[regex].match_count += 1
-            self.regex_extraction_stats[regex].bytes_matched += bytes_match.match_length
-            self.regex_extraction_stats[regex].bytes_match_objs.append(bytes_match)
+        yaralyzer = Yaralyzer.for_patterns([escape_yara_pattern(pattern)], self.bytes, self.label.plain, rules_label)
+
+        for bytes_match, bytes_decoder in yaralyzer.match_iterator():
+            self.regex_extraction_stats[pattern].match_count += 1
+            self.regex_extraction_stats[pattern].bytes_matched += bytes_match.match_length
+            self.regex_extraction_stats[pattern].bytes_match_objs.append(bytes_match)
 
             # Send suppressed decodes to a queue and track the reason for the suppression in the stats
-            if not (force or PdfalyzerConfig.MIN_DECODE_LENGTH < bytes_match.match_length < PdfalyzerConfig.MAX_DECODE_LENGTH):
-                self._queue_suppression_notice(bytes_match, label)
+            if not (force or (YaralyzerConfig.MIN_DECODE_LENGTH < bytes_match.match_length < YaralyzerConfig.MAX_DECODE_LENGTH)):
+                self._queue_suppression_notice(bytes_match, rules_label)
                 continue
 
             # Print out any queued suppressed notices before printing non suppressed matches
             self._print_suppression_notices()
-            self._attempt_binary_decodes(bytes_match, label or clean_byte_string(regex.pattern))
+            self._record_decode_stats(bytes_match, bytes_decoder, rules_label or pattern)
 
-        if self.regex_extraction_stats[regex].match_count == 0:
-            console.print(f"{regex.pattern} was not found for {label}...", style='dim')
+            if self.regex_extraction_stats[pattern].match_count == 0:
+                console.print(f"{pattern} was not found for {self.label}...", style='dim')
 
-    def _attempt_binary_decodes(self, bytes_match: BytesMatch, label: str) -> None:
+    def _record_decode_stats(self, bytes_match: BytesMatch, decoder: BytesDecoder, label: str) -> None:
         """Attempt to decode _bytes with all configured encodings and print a table of the results"""
-        decoder = BytesDecoder(bytes_match, label)
-        decoder.print_decode_attempts()
-        console.line()
-
         # Track stats on whether the bytes were decodable or not w/a given encoding
         self.regex_extraction_stats[bytes_match.label].matches_decoded += 1
 
@@ -195,11 +191,11 @@ class BinaryScanner:
         self.regex_extraction_stats[bytes_match.label].skipped_matches_lengths[bytes_match.match_length] += 1
         txt = bytes_match.__rich__()
 
-        if bytes_match.match_length < PdfalyzerConfig.MIN_DECODE_LENGTH:
+        if bytes_match.match_length < YaralyzerConfig.MIN_DECODE_LENGTH:
             txt = Text('Too little to actually attempt decode at ', style='grey') + txt
         else:
             txt.append(" is too large to decode ")
-            txt.append(f"(--max-decode-length is {PdfalyzerConfig.MAX_DECODE_LENGTH} bytes)", style='grey')
+            txt.append(f"(--max-decode-length is {YaralyzerConfig.MAX_DECODE_LENGTH} bytes)", style='grey')
 
         log.debug(Text('Queueing suppression notice: ') + txt)
         self.suppression_notice_queue.append(txt)