+++ pdfalyzer/helpers/number_helper.py
@@ -3,11 +3,10 @@ from typing import Union
 
 from rich.text import Text
 
-from pdfalyzer.helpers.rich_text_helper import prefix_with_plain_text_obj
-from pdfalyzer.util.logging import log
+from yaralyzer.helpers.rich_text_helper import prefix_with_plain_text_obj
 
 
-def is_divisible_by(n: Union[Number, int], divisor: int) -> bool:
+def is_divisible_by(n: int, divisor: int) -> bool:
     return divmod(n, divisor)[1] == 0