+++ pdfalyzer/helpers/pdf_object_helper.py
@@ -2,15 +2,15 @@
 Some methods to help with the direct manipulation/processing of PyPDF2's PdfObjects
 """
 from collections import namedtuple
+from typing import List
 
 from PyPDF2.generic import IndirectObject, PdfObject
 from rich.markup import escape
+from yaralyzer.util.logging import log
 
 from pdfalyzer.util.adobe_strings import DANGEROUS_PDF_KEYS
 from pdfalyzer.util.exceptions import PdfWalkError
-from pdfalyzer.util.logging import log
 from pdfalyzer.helpers.rich_text_helper import get_label_style
-
 
 # In the case of easy key/value pairs the reference_key and the reference_address are the same but
 # for more complicated references the reference_address will be the reference_key plus sub references.
@@ -23,7 +23,7 @@ PdfObjectRef = namedtuple('PdfObjectRef', ['reference_key', 'reference_address',
 SymlinkRepresentation = namedtuple('SymlinkRepresentation', ['text', 'style'])
 
 
-def get_references(obj: PdfObject, ref_key=None, ref_address=None) -> [PdfObjectRef]:
+def get_references(obj: PdfObject, ref_key=None, ref_address=None) -> List[PdfObjectRef]:
     """Return list of PdfObjectRefs"""
     if isinstance(obj, IndirectObject):
         if ref_key is None or ref_address is None: