+++ pdfalyzer/util/argument_parser.py
@@ -1,22 +1,19 @@
-import importlib.metadata
-import logging
 import sys
 from argparse import ArgumentDefaultsHelpFormatter, ArgumentParser
 from collections import namedtuple
 from functools import partial, update_wrapper
+from importlib.metadata import version
 from os import environ, getcwd, path
 from typing import List
 
 from rich_argparse import RichHelpFormatter
+from yaralyzer.helpers.rich_text_helper import console
+from yaralyzer.util.argument_parser import export, parser, parse_arguments as parse_yaralyzer_args
+from yaralyzer.util.logging import log, log_and_print, log_argparse_result, log_current_config, log_invocation
 
-from pdfalyzer.config import (DEFAULT_MIN_DECODE_LENGTH, DEFAULT_MAX_DECODE_LENGTH,
-     SURROUNDING_BYTES_LENGTH_DEFAULT, LOG_DIR_ENV_VAR, PdfalyzerConfig)
+from pdfalyzer.config import LOG_DIR_ENV_VAR, PdfalyzerConfig
 from pdfalyzer.detection.constants.binary_regexes import QUOTE_REGEXES
-from pdfalyzer.detection.encoding_detector import CONFIDENCE_SCORE_RANGE, EncodingDetector
 from pdfalyzer.helpers import rich_text_helper
-from pdfalyzer.helpers.file_helper import timestamp_for_filename
-from pdfalyzer.helpers.rich_text_helper import console, console_width_possibilities
-from pdfalyzer.util.logging import invocation_log, log, log_and_print, log_current_config
 
 
 # NamedTuple to keep our argument selection orderly
@@ -32,7 +29,6 @@ class ExplicitDefaultsHelpFormatter(ArgumentDefaultsHelpFormatter):
             return super()._get_help_string(action)
 
 
-ARGPARSE_LOG_FORMAT = '{0: >30}    {1: <17} {2: <}\n'
 ALL_FONTS_OPTION = -1
 
 DESCRIPTION = "Explore PDF's inner data structure with absurdly large and in depth visualizations. " + \
@@ -46,10 +42,19 @@ EPILOG = "Values for various config options can be set permanently by a .pdfalyz
          "environment variable is configured."
 
 
-# Positional args, version, help, etc
-parser = ArgumentParser(formatter_class=ExplicitDefaultsHelpFormatter, description=DESCRIPTION, epilog=EPILOG)
-parser.add_argument('--version', action='version', version=f"pdfalyzer {importlib.metadata.version('pdfalyzer')}")
-parser.add_argument('pdf', metavar='file_to_analyze.pdf', help='PDF file to process')
+# Positional args, version, help, etc. Note that we extend the yaralyzer's parser and export
+export.add_argument('-bin', '--extract-binary-streams',
+                    action='store_const',
+                    const='bin',
+                    help='extract all binary streams in the PDF to separate files (requires pdf-parser.py)')
+
+
+parser = ArgumentParser(
+    formatter_class=ExplicitDefaultsHelpFormatter,
+    description=DESCRIPTION,
+    epilog=EPILOG,
+    parents=[parser],  # Extend yaralyzer args
+    add_help=False)
 
 
 # Output section selection
@@ -82,162 +87,35 @@ select.add_argument('-f', '--font',
 select.add_argument('-y', '--yara', action='store_true',
                     help="scan the PDF and all its decoded/decrypted binary streams with YARA rules")
 
-
-# Fine tuning
-tuning = parser.add_argument_group(
-    'FINE TUNING',
-    "Tune various aspects of the analyses and visualizations to your needs. As an example setting " + \
-        "a low --max-decode-length (or suppressing brute force binary decode attempts altogether) can " + \
-        "dramatically improve run times and only rarely leads to a fatal lack of insight.")
-
-tuning.add_argument('--maximize-width', action='store_true',
-                    help="maximize the display width to fill the terminal")
-
-tuning.add_argument('--quote-type',
+select.add_argument('--quote-type',
                     help='scan binary data for quoted data of this type only or all types if not set',
                     choices=list(QUOTE_REGEXES.keys()))
-
-tuning.add_argument('--suppress-chardet', action='store_true',
-                    help="suppress the display of the full table of chardet's encoding likelihood scores")
-
-tuning.add_argument('--surrounding-bytes',
-                    help="number of bytes to display before and after suspicious strings in font binaries",
-                    default=SURROUNDING_BYTES_LENGTH_DEFAULT,
-                    metavar='BYTES',
-                    type=int)
-
-tuning.add_argument('--suppress-decodes', action='store_true',
-                    help='suppress decode attempts for quoted bytes found in font binaries')
-
-tuning.add_argument('--min-decode-length',
-                    help='suppress decode attempts for quoted byte sequences shorter than N',
-                    default=DEFAULT_MIN_DECODE_LENGTH,
-                    metavar='N',
-                    type=int)
-
-tuning.add_argument('--max-decode-length',
-                    help='suppress decode attempts for quoted byte sequences longer than N',
-                    default=DEFAULT_MAX_DECODE_LENGTH,
-                    metavar='N',
-                    type=int)
-
-tuning.add_argument('--force-display-threshold',
-                    help="chardet.detect() scores encodings from 0-100pct but encodings with scores below this number " + \
-                         "will not be displayed anywhere",
-                    default=EncodingDetector.force_display_threshold,
-                    metavar='PCT_CONFIDENCE',
-                    type=int,
-                    choices=CONFIDENCE_SCORE_RANGE)
-
-tuning.add_argument('--force-decode-threshold',
-                    help="extremely high (AKA 'above this number') PCT_CONFIDENCE scores from chardet.detect() " + \
-                         "as to the likelihood some binary data was written with a particular encoding will cause " + \
-                         "the pdfalyzer to do a force decode of that with that encoding. " + \
-                         "(chardet is a sophisticated libary; this is pdfalyzer's way of harnessing that intelligence)",
-                    default=EncodingDetector.force_decode_threshold,
-                    metavar='PCT_CONFIDENCE',
-                    type=int,
-                    choices=CONFIDENCE_SCORE_RANGE)
-
-
-# Export options
-export = parser.add_argument_group(
-    'FILE EXPORT',
-    "Multiselect. Choosing nothing is choosing nothing. Sends what you see on the screen to various file " + \
-        "formats in parallel. Writes files to the current directory if --output-dir is not provided. " + \
-        "Filenames are expansion of the PDF filename though you can use --file-prefix to make your " +
-        "filenames more unique and beautiful to their beholder.")
-
-export.add_argument('-bin', '--extract-binary-streams',
-                    action='store_const',
-                    const='bin',
-                    help='extract all binary streams in the PDF to separate files (requires pdf-parser.py)')
-
-export.add_argument('-svg', '--export-svg',
-                    action='store_const',
-                    const='svg',
-                    help='export analysis to SVG images')
-
-export.add_argument('-txt', '--export-txt',
-                    action='store_const',
-                    const='txt',
-                    help='export analysis to ANSI colored text files')
-
-export.add_argument('-html', '--export-html',
-                    action='store_const',
-                    const='html',
-                    help='export analysis to styled html files')
-
-export.add_argument('-out', '--output-dir',
-                    metavar='OUTPUT_DIR',
-                    help='write files to OUTPUT_DIR instead of current dir, does nothing if not exporting a file')
-
-export.add_argument('-pfx', '--file-prefix',
-                    metavar='PREFIX',
-                    help='optional string to use as the prefix for exported files of any kind')
-
-export.add_argument('-sfx', '--file-suffix',
-                    metavar='SUFFIX',
-                    help='optional string to use as the suffix for exported files of any kind')
-
-
-# Debugging
-debug = parser.add_argument_group(
-    'DEBUG',
-    'Debugging/interactive options.')
-
-debug.add_argument('-I', '--interact', action='store_true',
-                    help='drop into interactive python REPL when parsing is complete')
-
-debug.add_argument('-D', '--debug', action='store_true',
-                    help='show extremely verbose debug log output')
 
 
 # The Parsening Begins
 def parse_arguments():
     """Parse command line args. Most settings are communicated to the app by setting env vars"""
+    if '--version' in sys.argv:
+        print(f"pdfalyzer {version('pdfalyzer')}")
+        sys.exit()
+
     args = parser.parse_args()
-
-    if args.debug:
-        log.setLevel(logging.DEBUG)
-
-    _log_invocation()
-    args.invoked_at_str = timestamp_for_filename()
-
-    if args.maximize_width:
-        rich_text_helper.console.width = max(console_width_possibilities())
-
-    # Suppressing/limiting output
-    PdfalyzerConfig.MIN_DECODE_LENGTH = args.min_decode_length
-    PdfalyzerConfig.MAX_DECODE_LENGTH = args.max_decode_length
-    PdfalyzerConfig.NUM_SURROUNDING_BYTES = args.surrounding_bytes
+    args = parse_yaralyzer_args(args)
+    log_invocation()
 
     if args.quote_type:
         PdfalyzerConfig.QUOTE_TYPE = args.quote_type
-
-    if args.suppress_decodes:
-        PdfalyzerConfig.SUPPRESS_DECODES = args.suppress_decodes
-
-    # chardet.detect() action thresholds
-    if args.force_decode_threshold:
-        EncodingDetector.force_decode_threshold = args.force_decode_threshold
-
-    if args.force_display_threshold:
-        EncodingDetector.force_display_threshold = args.force_display_threshold
-
-    if args.suppress_chardet:
-        PdfalyzerConfig.SUPPRESS_CHARDET_OUTPUT = True
 
     # File export options
     if args.export_svg or args.export_txt or args.export_html or args.extract_binary_streams:
         args.output_dir = args.output_dir or getcwd()
         file_prefix = (args.file_prefix + '__') if args.file_prefix else ''
         args.file_suffix = ('_' + args.file_suffix) if args.file_suffix else ''
-        args.output_basename =  f"{file_prefix}{path.basename(args.pdf)}"
+        args.output_basename =  f"{file_prefix}{path.basename(args.file_to_scan_path)}"
     elif args.output_dir:
         log.warning('--output-dir provided but no export option was chosen')
 
-    _log_argparse_result(args)
+    log_argparse_result(args)
     log_current_config()
     return args
 
@@ -270,22 +148,3 @@ def output_sections(args, pdfalyzer) -> List[OutputSection]:
         return possible_output_sections
     else:
         return output_sections
-
-
-def _log_invocation() -> None:
-    """Log the command used to launch the pdfalyzer to the invocation log"""
-    invocation_log.info(f"THE INVOCATION: '{' '.join(sys.argv)}'")
-
-
-def _log_argparse_result(args):
-    """Logs the result of argparse"""
-    args_dict = vars(args)
-    log_msg = 'argparse results:\n' + ARGPARSE_LOG_FORMAT.format('OPTION', 'TYPE', 'VALUE')
-
-    for arg_var in sorted(args_dict.keys()):
-        arg_val = args_dict[arg_var]
-        row = ARGPARSE_LOG_FORMAT.format(arg_var, type(arg_val).__name__, str(arg_val))
-        log_msg += row
-
-    log_msg += "\n"
-    invocation_log.info(log_msg)