+++ pdfalyzer/util/debugging.py
@@ -1,5 +1,4 @@
 import logging
-from pdfalyzer.helpers.rich_text_helper import console
 
 
 # Starting pdb.set_trace() this way kind of sucks because yr locals are messed up